apps/inventory/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Inventory, InventoryLog

class InventorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Inventory
        fields = '__all__'

class InventoryLogSerializer(serializers.ModelSerializer):
    class Meta:
        model = InventoryLog
        fields = '__all__'
