requirements.txt
Django==5.0
djangorestframework
redis
kafka-python
psycopg2-binary
celery[redis]>=5.2
python-decouple
gunicorn>=20.1
channels>=4.0
channels-redis
djangorestframework-simplejwt
django-cors-headers
Pillow==12.0.0
# Websockets / ASGI
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
asgiref==3.8.0
razorpay
reportlab




